razzv4/RAZZv4-backend/pyproject.toml
[project]
name = "razzv4-backend"
version = "0.1.0"
description = "FastAPI backend for RAZZv4 application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.37",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.7.1",
    "jinja2>=3.1.5",
    "ultralytics>=8.3.0",
    "opencv-python-headless>=4.10.0.84",
    "numpy>=2.2.2",
    "pillow>=11.1.0",
    "websockets>=14.0",
    "bcrypt>=4.0.0",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.16.0",
    "pgvector>=0.2.0",
    "scipy>=1.11.0",
    "shapely>=2.0.0",
    "torchreid>=0.2.5",
    "faiss-cpu>=1.12.0",
    "gdown>=5.2.0",
    "tensorboard>=2.20.0",
]
